server/requirements.txt
amqp==2.6.1
billiard==3.6.4.0
celery==4.4.7
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
flower==0.9.7
greenlet==1.1.0
h5py==3.2.1
humanize==3.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
Keras==2.4.3
kombu==4.6.11
MarkupSafe==1.1.1
mtcnn==0.1.0
numpy==1.20.3
opencv-python==4.5.2.52
Pillow==8.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.18
psycopg2==2.8.6
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
scipy==1.6.3
six==1.16.0
SQLAlchemy==1.4.16
tornado==6.1
vine==1.3.0
wcwidth==0.2.5
Werkzeug==1.0.1